Reorganize unit tests
